Add metadata action to jib push / Merge build+push (#201)
I added the metadata action that will create image tags based on the
repository config, and will also push those tags to the respective
registries. It is assumes that the present docker daemon is current
authenticated against the registry you want to push to. This is usually
done docker-login action.

---------

Co-authored-by: kubectl <me@kube.moe>