bump: docker publish to 1.46.3 (#204)
The 1.16.0 publish action does assume that images that are loaded are
tagged in a certian way. This is not guaranteed anymore and the newer
versions address this (i.e. just using the image name without a tag to
retag, though I have no idea how docker internally chooses what image to
use. Should not be an issue cuz only one image tag will be loaded
anyway)

It also required the addition of the login action to facilitate auth
against registries

---------

Co-authored-by: kubectl <me@kube.moe>